surprise_trip/pyproject.toml
[tool.poetry]
name = "surprise_travel"
version = "0.1.0"
description = "surprise-travel using crewAI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = {extras = ["tools"], version = "^0.130.0"}
python-dotenv = "^1.0.0"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"

[tool.poetry.scripts]
surprise_travel = "surprise_travel.main:run"
train = "surprise_travel.main:train"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
